+++ src-native/library_types.rs
@@ -473,6 +473,7 @@ pub struct Playlist {
 		deserialize_with = "deserialize_playlist_ids",
 		serialize_with = "serialize_playlist_ids"
 	)]
+	#[cfg_attr(feature = "napi", napi(ts_type = "string[]"))]
 	pub tracks: Vec<ItemId>,
 }
 impl Playlist {